Fixes #131 : Added Filters in Media Editor
