+++ app/src/main/res/layout/gallery3d_cropimage.xml
@@ -191,6 +191,7 @@
                             android:id="@+id/btnRotateLeftRight"
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
+                            android:background="@null"
                             android:layout_alignParentLeft="true"
                             android:layout_alignParentTop="true"
                             android:contentDescription="@string/empty"
@@ -198,26 +199,31 @@
 
                         <ImageButton
                             android:id="@+id/btnSnow"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
+                            android:layout_width="100dp"
+                            android:scaleType="centerCrop"
+                            android:layout_height="100dp"
+                            android:background="@null"
                             android:layout_alignParentTop="true"
                             android:layout_toRightOf="@+id/btnSnow"
                             android:contentDescription="@string/empty"
-                            android:src="@drawable/snow" />
+                            android:src="@drawable/snow_filter" />
 
                         <ImageButton
                             android:id="@+id/btnRoundCorner"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
+                            android:layout_width="100dp"
+                            android:scaleType="centerCrop"
+                            android:layout_height="100dp"
+                            android:background="@null"
+                            android:src="@drawable/round_filter"
                             android:layout_alignParentTop="true"
                             android:layout_toRightOf="@+id/btnRoundCorner"
-                            android:contentDescription="@string/empty"
-                            android:src="@drawable/round_cornor" />
+                            android:contentDescription="@string/empty" />
 
                         <ImageButton
                             android:id="@+id/btnRotateTopDown"
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
+                            android:background="@null"
                             android:layout_alignParentRight="true"
                             android:layout_alignParentTop="true"
                             android:layout_toRightOf="@+id/btnOriginal"
